get_pipewire_state/status works
Finally found a reliable way to get the current status of pipewire so that the program can act accordingly. Also made use of "set_current_settings" in a way that can easier be reused.

When activating the radiobutton for suspending pipewire it will deactivate the other radiobuttons since they are no longer useable.

I'm sure there are many small things to iron out and things to improve.