Removing extra work
Split get_settings and show settings into different methods to call depending of what I want at the time.

Added a return tuple of 6 indexes to get_settings that are then used to check the current settings before setting the selected settings. for example. if you currently have Pipewire running at 64/48000 and press apply on 64/48000 if won't issue the commands since they are the same as the selected ones

I'm sure there are better ways to get the same result and I might look into them in the future but for now, it works enough to be usable